put directory sdk and grpc together under feature flags
