+++ libs/modkit/Cargo.toml
@@ -7,6 +7,7 @@ authors.workspace = true
 repository.workspace = true
 rust-version.workspace = true
 description = "Core ModKit library"
+readme.workspace = true
 
 [lib]
 name = "modkit"
@@ -31,7 +32,7 @@ otel = [
 bootstrap = [
     "dep:serde-saphyr",
     "dep:serde_yaml",
-    "dep:cf-system-directory-grpc",
+    "cf-system-sdks/directory_grpc",
     "dep:tracing-appender",
     "dep:file-rotate",
     "dep:tracing-log",
@@ -48,8 +49,7 @@ modkit-errors = { workspace = true, features = ["utoipa", "axum"] }
 modkit-db = { workspace = true }
 modkit-odata = { workspace = true, features = ["with-odata-params"] }
 modkit-sdk = { workspace = true }
-cf-system-sdks = { workspace = true } # TODO: probably we need remove this dependency at the future
-cf-system-directory-grpc = { workspace = true, optional = true }
+cf-system-sdks = { workspace = true, features = ["directory"] }
 
 # Core deps
 anyhow = { workspace = true }
