+++ libs/modkit/src/bootstrap/oop.rs
@@ -61,8 +61,7 @@ use crate::runtime::{
     ClientRegistration, DbOptions, MODKIT_DIRECTORY_ENDPOINT_ENV, RunOptions, ShutdownOptions, run,
     shutdown,
 };
-use cf_system_directory_grpc::DirectoryGrpcClient;
-use cf_system_sdks::directory::DirectoryClient;
+use cf_system_sdks::directory::{DirectoryClient, DirectoryGrpcClient};
 
 /// Configuration options for `OoP` module bootstrap
 #[derive(Debug, Clone)]
