+++ libs/system-sdks/Cargo.toml
@@ -14,6 +14,11 @@ name = "cf_system_sdks"
 [lints]
 workspace = true
 
+[features]
+directory = ["dep:cf-system-sdk-directory"]
+directory_grpc = ["directory", "cf-system-sdk-directory?/grpc"]
+
+
 [dependencies]
-cf-system-sdk-directory = { workspace = true }
+cf-system-sdk-directory = { workspace = true, optional = true }
