--- libs/system-sdks/directory-grpc/Cargo.toml
@@ -1,31 +1,0 @@
-[package]
-name = "cf-system-directory-grpc"
-version = "0.1.0"
-edition.workspace = true
-license.workspace = true
-authors.workspace = true
-description = "gRPC transport layer for directory service"
-build = "build.rs"
-include = ["proto"]
-
-[lib]
-name = "cf_system_directory_grpc"
-
-[dependencies]
-cf-system-sdks = { workspace = true }
-modkit-transport-grpc = { workspace = true }
-tonic = { workspace = true, features = ["transport"] }
-tonic-prost = { workspace = true }
-prost = { workspace = true }
-async-trait = { workspace = true }
-anyhow = { workspace = true }
-tracing = { workspace = true }
-
-[build-dependencies]
-tonic-prost-build = { workspace = true }
-
-[dev-dependencies]
-tokio = { workspace = true }
-
-[lints.rust]
-unexpected_cfgs = { level = "warn", check-cfg = ['cfg(coverage_nightly)'] }
