--- libs/system-sdks/directory-grpc/build.rs
@@ -1,11 +1,0 @@
-fn main() -> Result<(), Box<dyn std::error::Error>> {
-    println!("cargo:rerun-if-changed=proto/v1/directory.proto");
-    println!("cargo:rerun-if-changed=proto");
-
-    tonic_prost_build::configure()
-        .build_client(true)
-        .build_server(true)
-        .compile_protos(&["proto/v1/directory.proto"], &["proto"])?;
-
-    Ok(())
-}
