+++ libs/system-sdks/sdks/directory/Cargo.toml
@@ -7,6 +7,8 @@ authors.workspace = true
 repository.workspace = true
 description = "Directory service SDK module"
 rust-version.workspace = true
+build = "build.rs"
+include = ["proto", "src", "build.rs"]
 
 [lib]
 name = "cf_system_sdk_directory"
@@ -14,6 +16,29 @@ name = "cf_system_sdk_directory"
 [lints]
 workspace = true
 
+[features]
+grpc = [
+    "dep:modkit-transport-grpc",
+    "dep:tonic",
+    "dep:tonic-prost",
+    "dep:prost",
+    "dep:tracing",
+    "dep:tonic-prost-build",
+]
+
 [dependencies]
-async-trait = { workspace = true }
 anyhow = { workspace = true }
+async-trait = { workspace = true }
+
+
+modkit-transport-grpc = { workspace = true, optional = true }
+tonic = { workspace = true, features = ["transport"], optional = true }
+tonic-prost = { workspace = true, optional = true }
+prost = { workspace = true, optional = true }
+tracing = { workspace = true, optional = true }
+
+[build-dependencies]
+tonic-prost-build = { workspace = true, optional = true }
+
+[dev-dependencies]
+tokio = { workspace = true }
