+++ libs/system-sdks/sdks/directory/proto/v1/directory.proto
