+++ libs/system-sdks/src/lib.rs
@@ -1,1 +1,2 @@
+#[cfg(feature = "directory")]
 pub use cf_system_sdk_directory as directory;
