+++ modules/system/module_orchestrator/src/lib.rs
@@ -12,5 +12,6 @@ pub use module::{ModuleOrchestrator, ModuleOrchestratorConfig};
 mod server;
 
 // === RE-EXPORTS ===
-pub use cf_system_directory_grpc::DirectoryGrpcClient;
-pub use cf_system_sdks::directory::{RegisterInstanceInfo, ServiceEndpoint, ServiceInstanceInfo};
+pub use cf_system_sdks::directory::{
+    DirectoryGrpcClient, RegisterInstanceInfo, ServiceEndpoint, ServiceInstanceInfo,
+};
