+++ proto/directory
@@ -1,1 +1,1 @@
-../libs/system-sdks/directory-grpc/proto
+../libs/system-sdks/sdks/directory/proto
